Project/src/testcases/InternetTest.java
package testcases;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;
import org.junit.Test;
import classes.Internet;
import net.datafaker.Faker;

public class InternetTest {
	Faker faker = new Faker();
	
	@Test
    public void testdomainname()
    {
		String input = faker.internet().domainName();
		//System.out.println(input);
        assertTrue(Internet.verifydomainname(input));
    }
	
	@Test
    public void testdomainnsuffix()
    {
		String input = faker.internet().domainSuffix();
		//System.out.println(input);
        assertTrue(Internet.verifystring(input));
    }
	
	@Test
    public void testdomainnword()
    {
		String input = faker.internet().domainWord();
		//System.out.println(input);
		assertTrue(Internet.verifystring(input));
    }
	
	@Test
    public void testemail()
    {
		String input = faker.internet().emailAddress();
		//System.out.println(input);
		assertTrue(Internet.verifyemail(input));
    }
	
	@Test
    public void testhttp()
    {
		String input = faker.internet().httpMethod();
		//System.out.println(input);
		assertTrue(Internet.verifystring(input));
    }
	
	@Test
    public void testemailname()
    {
		String input = faker.internet().emailAddress("jeffrey");
		//System.out.println(input);
		assertTrue(Internet.verifyemail(input));
    }
	
	@Test
    public void testimage()
    {
		String input = faker.internet().image();
		String input2 = faker.internet().image(500, 500);
		String input3 = faker.internet().image(500, 500, "tree");
		assertTrue(Internet.verifydomainname(input));
		assertTrue(Internet.verifydomainname(input2));
		assertTrue(Internet.verifydomainname(input3));
    }
	
	@Test
    public void testip4()
    {
		String input = faker.internet().ipV4Address();
		//System.out.println(input);
		assertTrue(Internet.verifyip4(input));
    }
	
	@Test
    public void testip4cidr()
    {
		String input = faker.internet().ipV4Cidr();
		//System.out.println(input);
		assertTrue(Internet.verifyip4cidr(input));
    }
	
	@Test
    public void testip46()
    {
		String input = faker.internet().ipV6Address();
		//System.out.println(input);
		assertTrue(Internet.verifyipv6(input));
    }
	
	@Test
    public void testip46cidr()
    {
		String input = faker.internet().ipV6Cidr();
		//System.out.println(input);
		assertTrue(Internet.verifyipv6cidr(input));
    }
	
	@Test
    public void testmac()
    {
		String input = faker.internet().macAddress();
		//System.out.println(input);
		assertTrue(Internet.verifymac(input));
    }
	
	@Test
    public void testmacname()
    {
		String input = faker.internet().macAddress("Kenji");
		//System.out.println(input);
		assertTrue(Internet.verifymacname(input));
    }
	
	@Test
    public void testpassword()
    {
		String input = faker.internet().password();
		//System.out.println(input);
		assertTrue(Internet.verifystring(input));
    }
	
	@Test
    public void testpasswordtypes1()
    {
		String input = faker.internet().password(true);
		//System.out.println(input);
		assertTrue(Internet.verifypassnum(input));
    }
	
	@Test
    public void testpasswordtypes2()
    {
		String input2 = faker.internet().password(0, 5);
		//System.out.println(input2);
		assertTrue(Internet.verifypassnum2(input2, 0, 5));
    }
	
	@Test
    public void testpasswordtypes3()
    {
		String input3 = faker.internet().password(0, 5, true);
		//System.out.println(input3);
		assertTrue(Internet.verifypassnum3(input3, 0, 5));
    }
	
	@Test
    public void testpasswordtypes4() //THIS DATA HAS PROBLEM!!
    //USE THIS TEST
    {
		String input4 = faker.internet().password(0, 5, true, true, true);
		//input4 sometimes generates as a null string with no data
		//System.out.println(input4);
		//assertTrue(Internet.verifypassnum4(input4, 0, 5));
    }
	
	@Test
	public void testpasswordCACC()
	{
		String PA16 = faker.internet().password(5, 0, false, false, false); //Prints < 4 ERROR!
		String PB24 = faker.internet().password(0, 5, false, false, false); //Prints < 5 and > 0 as expected!
		String PC28 = faker.internet().password(0, 0, true, false, false); // Prints nothing as expected!
		String PD30 = faker.internet().password(0, 0, false, true, false); // Prints nothing as expected!
		String PE31 = faker.internet().password(0, 0, false, false, true); // Prints nothing as expected!
		String PABCDE1 = faker.internet().password(1, 10, true, true, true); //Adding one more test for CACC
		String PABCDE32 = faker.internet().password(0, 0, false, false, false); //Prints nothing as expected!
		
		System.out.println(PA16);
		assertFalse(Internet.verifypassnum4(PA16, 5, 0, false, false , false)); //Fails when it should return true!
		assertTrue(Internet.verifypassnum4(PB24, 0, 5, false, false , false));
		assertTrue(Internet.verifypassnum4(PC28, 0, 0, true, false , false));
		assertTrue(Internet.verifypassnum4(PD30, 0, 0, false, true , false));
		assertTrue(Internet.verifypassnum4(PE31, 0, 0, false, false , true));
		assertTrue(Internet.verifypassnum4(PABCDE1, 1, 10, true, true, true));
		assertTrue(Internet.verifypassnum4(PABCDE32, 0, 0, false, false , false));
	}
	
	@Test
    public void testprivip4()
    {
		String input = faker.internet().privateIpV4Address();
		//System.out.println(input);
		assertTrue(Internet.verifyip4(input));
    }
	
	@Test
    public void testpubicp4()
    {
		String input = faker.internet().publicIpV4Address();
		//System.out.println(input);
		assertTrue(Internet.verifyip4(input));
    }
	
	@Test
    public void testsafeemail()
    {
		String input = faker.internet().safeEmailAddress();
		//System.out.println(input);
		assertTrue(Internet.verifyemail(input));
    }
	
	@Test
    public void testslug()
    {
		String input = faker.internet().slug();
		//System.out.println(input);
		assertTrue(Internet.verifystring(input));
    }
	
	@Test
    public void testurl()
    {
		String input = faker.internet().url();
		//System.out.println(input);
		assertTrue(Internet.verifyurl(input));
    }
	
	@Test
    public void testuseragent()
    {
		String input = faker.internet().userAgent();
		//System.out.println(input);
		assertTrue(Internet.verifystring(input));
    }
	
	@Test
    public void testuserid()
    {
		String input = faker.internet().uuid();
		//System.out.println(input);
		assertTrue(Internet.verifyuid(input));
    }
	
	@Test
    public void testuserid3()
    {
		String input = faker.internet().uuidv3();
		//System.out.println(input);
		assertTrue(Internet.verifyuid(input));
    }
	
	@Test
    public void testport()
    {
		int input = faker.internet().port();
		assertTrue(Internet.verifyport(input));
    }
	
}
